+++ inlang/source-code/paraglide/paraglide-js-adapter-sveltekit/src/preprocessor/index.ts
@@ -1,17 +1,8 @@
 import { parse } from "svelte/compiler"
 import MagicString from "magic-string"
-import {
-	HEADER_COMPONENT_MODULE_ID,
-	LANGUAGE_TAG_ALIAS,
-	OUTDIR_ALIAS,
-	TRANSLATE_PATH_FUNCTION_NAME,
-	TRANSLATE_PATH_MODULE_ID,
-} from "../constants.js"
 import { InjectHeader } from "./injectHeader.js"
 import { RewriteHrefs } from "./rewriteHrefs.js"
-
-export type Ast = ReturnType<typeof parse>
-export type TemplateNode = Ast["html"]
+import type { Ast } from "./types.js"
 
 type PreprocessorArgs = {
 	filename: string
@@ -37,21 +28,21 @@ export type PreprocessingPass = {
 }
 
 const PASSES: PreprocessingPass[] = [InjectHeader, RewriteHrefs]
-
 export function preprocess() {
 	return {
 		name: "@inlang/paraglide-js-adapter-sveltekit",
 		markup: ({ filename, content }: { content: string; filename: string }) => {
-			//dont' process built in Svelte files
+			const NOOP = { code: content }
+
+			//dont' process files generated by the framework
 			if (filename.includes(".svelte-kit")) {
-				return { code: content }
+				return NOOP
 			}
 
 			const passMask = PASSES.map((pass) => pass.condition({ filename, content }))
-			const shouldProcess = passMask.some((value) => value)
-
-			if (!shouldProcess) {
-				return { code: content }
+			const skipProcessing = passMask.every((pass) => pass === false)
+			if (skipProcessing) {
+				return NOOP
 			}
 
 			const code = new MagicString(content)
